LZ4 moved to Pipeline
