+++ MonoGame.Framework.Content.Pipeline/Utilities/LZ4/LZ4Codec.Unsafe64HC.Dirty.cs
@@ -64,7 +64,7 @@ IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 // ReSharper disable InconsistentNaming
 
-namespace LZ4n
+namespace Microsoft.Xna.Framework.Content.Pipeline.LZ4
 {
 	public static partial class LZ4Codec
 	{
